requirements.txt
google-cloud-bigquery==2.7.0
psycopg2>=2.8.6
pandas>=1.3.4
numpy>=1.22.1
google-auth==1.35.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.6
google-api-python-client==2.27.0
configparser
datetime
pyarrow